Create database anilze class. Add column market to positions in database. Edit function_when_model_work function in functions and fake_position_robot in main. Alpha 0.0021

+++ app/model_generator.py
@@ -533,6 +533,9 @@ def generate_my_models(
     i = 0
     times = []
 
+    if delete_old_models == "no":
+        return ""
+
     if delete_old_models:
         delete_model(f"{catalog}\\models\\", '')
